pyproject.toml
[tool.poetry]
name = "stock-analyst"
version = "0.1.0"
description = "Equity scoring engine — rule-based multi-block stock analysis pipeline"
authors = []
package-mode = false

[tool.poetry.dependencies]
python     = "^3.11"
yfinance   = "^0.2.36"
pandas     = "^2.0"
numpy      = "^2.0"
python-dotenv = "^1.0"      # load .env into os.environ at startup

[tool.poetry.group.dev.dependencies]
pytest     = "^8.0"
pytest-cov = "^5.0"
ruff       = "^0.4"         # linter + formatter

[tool.poetry.scripts]
analyse = "main:main"

[build-system]
requires      = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

# ---------------------------------------------------------------------------
# Ruff config (linter)
# ---------------------------------------------------------------------------
[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I"]    # pycodestyle errors, pyflakes, isort
